requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
debugpy==1.6.7.post1
decorator==5.1.1
Django==4.2.6
djangorestframework==3.14.0
executing==1.2.0
flatbuffers==23.5.26
fonttools==4.43.0
frozenlist==1.4.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
multidict==6.0.4
nest-asyncio==1.5.7
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.6
Werkzeug==3.0.0
wrapt==1.14.1
yarl==1.9.2
